# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/users/me/study-plans/new": {
      "post": {
        "tags": [
          "Study plan"
        ],
        "operationId": "addNewStudyPlan",
        "parameters": [
          {
            "name": "User-External-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewStudyPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudyPlanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/study-plan/{studyPlanExternalId}/semesters/{semesterType}": {
      "post": {
        "tags": [
          "Study plan"
        ],
        "operationId": "addSemester",
        "parameters": [
          {
            "name": "studyPlanExternalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "semesterType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SPRING",
                "AUTUMN"
              ]
            }
          },
          {
            "name": "User-External-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StudyPlanResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getCurrentUser",
        "parameters": [
          {
            "name": "User-External-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/study-plans": {
      "get": {
        "tags": [
          "Study plan"
        ],
        "operationId": "getStudyPlans",
        "parameters": [
          {
            "name": "User-External-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudyPlanResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "creationDate",
          "externalId",
          "name"
        ]
      },
      "CreateNewStudyPlanRequest": {
        "type": "object",
        "properties": {
          "curriculumVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "curriculumId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CurriculumResponse": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "studyLevel": {
            "type": "string",
            "enum": [
              "BACHELOR",
              "MASTER",
              "INTEGRATED",
              "DOCTOR"
            ]
          },
          "credits": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "creationDate",
          "credits",
          "externalId",
          "studyLevel",
          "title"
        ]
      },
      "StudyPlanResponse": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "completedCredits": {
            "type": "number",
            "format": "double"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "curriculum": {
            "$ref": "#/components/schemas/CurriculumResponse"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "completedCredits",
          "creationDate",
          "curriculum",
          "externalId",
          "startDate"
        ]
      },
...
    }
  }
}